requirements.txt
accelerate==0.22.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
anyio==3.7.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.2.0
bitsandbytes==0.41.1
certifi==2024.7.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.12
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
diskcache==5.6.3
einops==0.6.1
fastapi==0.99.1
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.12
langchain-community==0.2.10
langchain-core==0.2.27
langchain-text-splitters==0.2.2
langsmith==0.1.96
llama_cpp_python==0.2.85
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
nltk==3.8.1
numexpr==2.10.1
numpy==1.26.4
onnxruntime==1.18.1
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pillow==10.4.0
posthog==3.5.0
protobuf==5.27.3
psutil==6.0.0
pulsar-client==3.5.0
pydantic==1.10.17
Pygments==2.18.0
PyPika==0.48.9
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==2.2.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.27.0
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.4
transformers==4.30.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.4
watchfiles==0.22.0
websockets==12.0
xformers==0.0.21
yarl==1.9.4
